+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Anticipated release: February 18, 2020
 
 #### 🚀 New features
 
+- Activity state personnel now show calculated total cost instead of just rate and FTEs. ([#2013])
+
 #### 🐛 Bugs fixed
 
 - Fixed a bug where autosave happened too often, causing performance issues, and potentially inconsistent saved data.
@@ -13,3 +15,5 @@ Anticipated release: February 18, 2020
 # Previous releases
 
 See our [release history](https://github.com/18F/cms-hitech-apd/releases)
+
+[#2013]: https://github.com/18F/cms-hitech-apd/issues/2013